readem/requirements.txt
PyMuPDF
bert-extractive-summarizer
flask
pyttsx3